+++ conn_pool.go
@@ -22,6 +22,7 @@ type ConnPool struct {
 	logger               Logger
 	logLevel             int
 	closed               bool
+	preparedStatements   map[string]*PreparedStatement
 }
 
 type ConnPoolStat struct {
@@ -58,6 +59,7 @@ func NewConnPool(config ConnPoolConfig) (p *ConnPool, err error) {
 
 	p.allConnections = make([]*Conn, 0, p.maxConnections)
 	p.availableConnections = make([]*Conn, 0, p.maxConnections)
+	p.preparedStatements = make(map[string]*PreparedStatement)
 	p.cond = sync.NewCond(new(sync.Mutex))
 
 	// Initially establish one connection
@@ -193,6 +195,19 @@ func (p *ConnPool) Reset() {
 	p.availableConnections = make([]*Conn, 0, p.maxConnections)
 }
 
+// invalidateAcquired causes all acquired connections to be closed when released.
+// The pool must already be locked.
+func (p *ConnPool) invalidateAcquired() {
+	p.resetCount++
+
+	for _, c := range p.availableConnections {
+		c.poolResetCount = p.resetCount
+	}
+
+	p.allConnections = p.allConnections[:len(p.availableConnections)]
+	copy(p.allConnections, p.availableConnections)
+}
+
 // Stat returns connection pool statistics
 func (p *ConnPool) Stat() (s ConnPoolStat) {
 	p.cond.L.Lock()
@@ -204,18 +219,28 @@ func (p *ConnPool) Stat() (s ConnPoolStat) {
 	return
 }
 
-func (p *ConnPool) createConnection() (c *Conn, err error) {
-	c, err = Connect(p.config)
+func (p *ConnPool) createConnection() (*Conn, error) {
+	c, err := Connect(p.config)
 	if err != nil {
-		return
+		return nil, err
 	}
+
 	if p.afterConnect != nil {
 		err = p.afterConnect(c)
 		if err != nil {
-			return
+			c.die(err)
+			return nil, err
 		}
 	}
-	return
+
+	for _, ps := range p.preparedStatements {
+		if _, err := c.Prepare(ps.Name, ps.SQL); err != nil {
+			c.die(err)
+			return nil, err
+		}
+	}
+
+	return c, nil
 }
 
 // Exec acquires a connection, delegates the call to that connection, and releases the connection
@@ -261,6 +286,64 @@ func (p *ConnPool) QueryRow(sql string, args ...interface{}) *Row {
 // transaction is closed the connection will be automatically released.
 func (p *ConnPool) Begin() (*Tx, error) {
 	return p.BeginIso("")
+}
+
+// Prepare creates a prepared statement on a connection in the pool to test the
+// statement is valid. If it succeeds all connections accessed through the pool
+// will have the statement available.
+//
+// Prepare creates a prepared statement with name and sql. sql can contain
+// placeholders for bound parameters. These placeholders are referenced
+// positional as $1, $2, etc.
+//
+// Prepare is idempotent; i.e. it is safe to call Prepare multiple times with
+// the same name and sql arguments. This allows a code path to Prepare and
+// Query/Exec without concern for if the statement has already been prepared.
+func (p *ConnPool) Prepare(name, sql string) (*PreparedStatement, error) {
+	p.cond.L.Lock()
+	defer p.cond.L.Unlock()
+
+	if ps, ok := p.preparedStatements[name]; ok && ps.SQL == sql {
+		return ps, nil
+	}
+
+	c, err := p.acquire()
+	if err != nil {
+		return nil, err
+	}
+	ps, err := c.Prepare(name, sql)
+	p.availableConnections = append(p.availableConnections, c)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, c := range p.availableConnections {
+		_, err := c.Prepare(name, sql)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	p.invalidateAcquired()
+	p.preparedStatements[name] = ps
+
+	return ps, err
+}
+
+// Deallocate releases a prepared statement from all connections in the pool.
+func (p *ConnPool) Deallocate(name string) (err error) {
+	p.cond.L.Lock()
+	defer p.cond.L.Unlock()
+
+	for _, c := range p.availableConnections {
+		if err := c.Deallocate(name); err != nil {
+			return err
+		}
+	}
+
+	p.invalidateAcquired()
+
+	return nil
 }
 
 // BeginIso acquires a connection and begins a transaction in isolation mode iso